requirements.txt
asgiref==3.4.1
Django==3.2.5
djangorestframework==3.10.2
drf-nested-routers==0.91

